+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Restapi/Cli: Use pydantic models for request parameters
 - Rename `dropna` to `drop_nulls`
 - Change default of `drop_nulls` to True
+- Replace occurrences of `dt.timezone.utc` by `ZoneInfo("UTC")`
 
 ## 0.98.0 - 2024-12-09